RoutingService delegates to both Graph and GraphIndex
