+++ src/main/java/org/opentripplanner/routing/graph/Graph.java
@@ -13,7 +13,6 @@ import com.google.common.collect.Maps;
 import com.google.common.collect.Multimap;
 import gnu.trove.list.TDoubleList;
 import gnu.trove.list.linked.TDoubleLinkedList;
-import lombok.experimental.Delegate;
 import org.apache.commons.math3.stat.descriptive.rank.Median;
 import org.joda.time.DateTime;
 import org.locationtech.jts.geom.Coordinate;
@@ -28,7 +27,6 @@ import org.opentripplanner.ext.siri.updater.SiriSXUpdater;
 import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.graph_builder.issues.NoFutureDates;
 import org.opentripplanner.model.Agency;
-import org.opentripplanner.model.Route;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.calendar.CalendarService;
 import org.opentripplanner.model.FeedInfo;
@@ -1065,97 +1063,9 @@ public class Graph implements Serializable {
         return getService(BikeRentalStationService.class);
     }
 
-    public Map<String, Map<String, Agency>> getAgenciesForFeedId() {
-        return index.getAgenciesForFeedId();
-    }
-
-    public Map<FeedScopedId, Operator> getOperatorForId() {
-        return index.getOperatorForId();
-    }
-
-    public Map<String, FeedInfo> getFeedInfoForId() {
-        return index.getFeedInfoForId();
-    }
-
-    public Map<FeedScopedId, Stop> getStopForId() {
-        return index.getStopForId();
-    }
-
-    public Map<FeedScopedId, Trip> getTripForId() {
-        return index.getTripForId();
-    }
-
-    public Map<FeedScopedId, Route> getRouteForId() {
-        return index.getRouteForId();
-    }
-
-    public Map<Stop, TransitStopVertex> getStopVertexForStop() {
-        return index.getStopVertexForStop();
-    }
-
-    public Map<Trip, TripPattern> getPatternForTrip() {
-        return index.getPatternForTrip();
-    }
-
-    public Multimap<Route, TripPattern> getPatternsForRoute() {
-        return index.getPatternsForRoute();
-    }
-
-    public Multimap<Stop, TripPattern> getPatternsForStop() {
-        return index.getPatternsForStop();
-    }
-
-    public Map<Station, MultiModalStation> getMultiModalStationForStations() {
-        return index.getMultiModalStationForStations();
-    }
-
-    public Set<Route> getRoutesForStop(Stop stop) {
-        return index.getRoutesForStop(stop);
-    }
-
-    /**
-     * Fetch an agency by its string ID, ignoring the fact that this ID should be scoped by a feedId.
-     * This is a stopgap (i.e. hack) method for fetching agencies where no feed scope is available. I
-     * am creating this method only to allow merging pull request #2032 which adds GraphQL. Note that
-     * if the same agency ID is defined in several feeds, this will return one of them at random. That
-     * is obviously not the right behavior. The problem is that agencies are not currently keyed on an
-     * FeedScopedId object, but on separate feedId and id Strings. A real fix will involve replacing
-     * or heavily modifying the OBA GTFS loader, which is now possible since we have forked it.
-     */
-    public Agency getAgencyWithoutFeedId(String agencyId) {
-        // Iterate over the agency map for each feed.
-        for (Map<String, Agency> agencyForId : getAgenciesForFeedId().values()) {
-            Agency agency = agencyForId.get(agencyId);
-            if (agency != null) {
-                return agency;
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Construct a set of all Agencies in this graph, spanning across all feed IDs. I am creating this
-     * method only to allow merging pull request #2032 which adds GraphQL. This should probably be
-     * done some other way, see javadoc on getAgencyWithoutFeedId.
-     */
-    public Set<Agency> getAllAgencies() {
-        Set<Agency> allAgencies = new HashSet<>();
-        for (Map<String, Agency> agencyForId : getAgenciesForFeedId().values()) {
-            allAgencies.addAll(agencyForId.values());
-        }
-        return allAgencies;
-    }
-
     public Collection<Notice> getNoticesByEntity(TransitEntity<?> entity) {
         Collection<Notice> res = getNoticesByElement().get(entity);
         return res == null ? Collections.emptyList() : res;
-    }
-
-    /**
-     * Get a list of all operators spanning across all feeds.
-     */
-    public Collection<Operator> getAllOperators() {
-        return getOperatorForId().values();
     }
 
     public TripPattern getTripPatternForId(String id) {
